Move events to separate package
