+++ render-remote/src/commonMain/kotlin/ink/ui/render/remote/WebForwarder.kt
@@ -2,9 +2,9 @@ package ink.ui.render.remote
 
 import ink.ui.render.remote.serialization.ElementId
 import ink.ui.render.remote.serialization.LayoutMessage
-import ink.ui.render.remote.serialization.OnClickEvent
-import ink.ui.render.remote.serialization.UiEvent
-import ink.ui.render.remote.serialization.UiEvents
+import ink.ui.render.remote.serialization.event.OnClickEvent
+import ink.ui.render.remote.serialization.event.UiEvent
+import ink.ui.render.remote.serialization.event.UiEvents
 import ink.ui.structures.elements.ButtonElement
 import ink.ui.structures.elements.UiElement
 import ink.ui.structures.layouts.UiLayout
