+++ render-remote/src/commonMain/kotlin/ink/ui/render/remote/serialization/event/UiEvent.kt
@@ -1,4 +1,6 @@
-package ink.ui.render.remote.serialization
+package ink.ui.render.remote.serialization.event
+
+import ink.ui.render.remote.serialization.ElementId
 
 /**
  * Associates a UI event with a specific occurrence of an event.
